@claude fix(arch): move PhotobankRepository to Persistence layer (ADR-004)
Relocates PhotobankRepository from Anela.Heblo.Application/Features/Photobank/
to Anela.Heblo.Persistence/Photobank/, aligning it with every other DB-backed
repository in the codebase and eliminating the lone ADR-004 violation.

- Moves the concrete class to Persistence/Photobank/ with namespace
  Anela.Heblo.Persistence.Photobank (file-scoped, matching EF config neighbors)
- Updates PhotobankModule.cs: replaces `using Anela.Heblo.Persistence` and
  `using Microsoft.EntityFrameworkCore` (both unused in the module) with
  `using Anela.Heblo.Persistence.Photobank`
- Fixes 6 test files that instantiate PhotobankRepository directly: replaces
  `using Anela.Heblo.Application.Features.Photobank` with
  `using Anela.Heblo.Persistence.Photobank`
- Zero logic changes; all 5148 unit tests pass; dotnet build 0 errors;
  dotnet format --verify-no-changes exits 0
