+++ backend/src/Anela.Heblo.Application/Features/Photobank/PhotobankModule.cs
@@ -16,9 +16,8 @@ using Anela.Heblo.Application.Features.Photobank.UseCases.UpdateRule;
 using Anela.Heblo.Application.Features.Photobank.Validators;
 using Anela.Heblo.Domain.Features.Configuration;
 using Anela.Heblo.Domain.Features.Photobank;
-using Anela.Heblo.Persistence;
+using Anela.Heblo.Persistence.Photobank;
 using FluentValidation;
-using Microsoft.EntityFrameworkCore;
 using MediatR;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
