+++ backend/test/Anela.Heblo.Tests/Features/Photobank/PhotobankRepositoryFilterTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using Anela.Heblo.Domain.Features.Photobank;
 using Anela.Heblo.Persistence;
-using Anela.Heblo.Application.Features.Photobank;
+using Anela.Heblo.Persistence.Photobank;
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using Xunit;
