+++ backend/test/Anela.Heblo.Tests/Features/Photobank/PhotobankRepositoryGetTagsTests.cs
@@ -2,7 +2,7 @@ using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Anela.Heblo.Application.Features.Photobank;
+using Anela.Heblo.Persistence.Photobank;
 using Anela.Heblo.Domain.Features.Photobank;
 using Anela.Heblo.Persistence;
 using FluentAssertions;
